Modified cnt type to dou

+++ GitDemoConflits/src/conflictDemo/Fibonacci.java
@@ -8,7 +8,7 @@ public class Fibonacci {
 
 		int n1 = 0, n2 = 1, n3 = n1 + n2;
 		int upto = 8;
-		int cnt = 3;
+		double cnt = 3;
 		System.out.println(n1 + " " + n2 + " " + n3 + " ");
 	
 		while (cnt < upto) {
